refactor: 파일 구조를 기능에 맞게 변경
