+++ src/main/java/chess/domain/board/ChessBoard.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import chess.domain.piece.Bishop;
 import chess.domain.piece.Blank;
@@ -6,10 +6,12 @@ import chess.domain.piece.King;
 import chess.domain.piece.Knight;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
+import chess.domain.piece.Pieces;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
-import chess.domain.player.Pieces;
-import chess.domain.player.Score;
+import chess.domain.piece.TeamColor;
+import chess.domain.result.Result;
+import chess.domain.result.Score;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
