+++ src/main/java/chess/domain/piece/TeamColor.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 public enum TeamColor {
     BLACK, WHITE, NONE
